Move: 브랜드 리스트 조회 기능 MemberController로 이동
